Remove `json_max` type
The `json_max` type is the type which includes all polymorphic variants.
This is done so the `Pretty` module can be reused through all the
variants, but it is somewhat obscure how that works since it requires
upcasting. This commit changes it to generate different `Pretty`
modules, each specialized to the exact selected variant.